# common/src/main/java/fr/epita/assistants/common/api/response/ItemResponse.java
package fr.epita.assistants.common.api.response;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class ItemResponse {
    private Integer id;
    private String type;
    private float quantity;
}