# common/src/main/java/fr/epita/assistants/common/api/response/ShopPriceResponse.java
package fr.epita.assistants.common.api.response;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class ShopPriceResponse {
    private float shopPrice;
}