# shop/src/main/resources/openapi.yaml
...
openapi: 3.1.0
...
components:
  schemas:
...
    ItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        quantity:
          type: number
          format: float
        type:
          $ref: "#/components/schemas/ResourceType"
...
    ItemsResponse:
      type: object
      properties:
        itemsResponse:
          type: array
          items:
            $ref: "#/components/schemas/ItemResponse"
    ResourceType:
      type: string
      enum:
      - MONEY
      - GROUND
      - WATER
      - ROCK
      - WOOD
    ShopPriceResponse:
      type: object
      properties:
        shopPrice:
          type: number
          format: float
...
paths:
...
  /price:
    get:
      summary: Get the current shop price.
      description: Retrieves the current shop price.
      x-quarkus-openapi-method-ref: m869844210_-271831484
      tags:
      - Shop Management
      responses:
        "200":
          description: The shop price was successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopPriceResponse"
        "400":
          description: The game is not running.
      servers:
      - url: http://localhost:8082/
  /resources:
    get:
      summary: Retrieve available resources in inventory.
      description: Fetch all resources currently available in the inventory.
      x-quarkus-openapi-method-ref: m911209932_-1184004262
      tags:
      - Inventory Management
      responses:
        "200":
          description: The resources were successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsResponse"
        "400":
          description: The game is not running or the request is invalid.
      servers:
      - url: http://localhost:8082/
...
info:
  title: shop API
  version: 1.0.0